refactor: Moved slice test folder, updated dependencies
