+++ tests/state/ViewerState/backdrop_slice.test.ts
@@ -5,7 +5,7 @@ import { Dataset } from "../../../src/colorizer";
 import { ANY_ERROR } from "../../test_utils";
 import { MOCK_DATASET_WITHOUT_BACKDROP } from "./constants";
 
-import { useViewerStateStore } from "../../../src/colorizer/state/ViewerState";
+import { useViewerStateStore } from "../../../src/state/ViewerState";
 
 describe("useViewerStateStore: BackdropSlice", () => {
   describe("setBackdropKey", () => {
